requirements-airflow.txt
apache-airflow[postgres]==1.10.8
pandas==1.0.0
pendulum==1.4.4
pyarrow==0.16.0
pyspark==2.4.3
python-snappy==0.5.4
werkzeug==0.16.1